+++ src/ReleaseNotes.md
@@ -2,7 +2,7 @@
 Minor release
 
 ## Features
-- Retargeting to support Windows 11 24H2 and newer only
+- Added support for Entra ID Managed Identity authentication
 
 ## Bug Fixes
-- Fixed options mismatch in the command line interface
+- None